Support attribute access for Promise.
This implements basic support for accessing attributes including methods on a
`Promise` object. The magic methods for several operators implemented in the
proxy class of `lazy` are also supported, so that type checking is
available for +, <, % on `Promise` objects.
https://github.com/django/django/blob/eb0f298e765ced5d52461ee8ec3fed8278b3d8d6/django/utils/functional.py#L166-L183

This plugin feature is implemented with `get_method_signature_hook` for
the magic operators since we need to override the method signature
defined on the stubs for the previously mentioned magic methods; and
`get_attribute_hook` to dynamically resolve the resultclasses to check
the type of the attribute. We might want to optimize this by caching
the resultclasses or attribute types.

Signed-off-by: Zixuan James Li <p359101898@gmail.com>